chore[python]: add debug messages and exceptions

+++ database.py
@@ -19,13 +19,25 @@ for o in os.listdir("./json"):
 # TODO make CRIT value based on PLAYERLVL
 
 # define function to remove a save slot
-def deleteSave(n):
+def deleteSave(n, rm):
     try:
         client.drop_database(n)
     except:
         print("Database doesn't exist")
-    if os.path.exists(savesPath + "/" + n):
-        shutil.rmtree(savesPath + "/" + n)
+        return
+    dbList = list(client.list_database_names())
+    if name in dbList:
+        raise Exception("MongoDB Error: '" + n + "' didn't get dropped.")
+        return
+    if eng.DEBUG == 1:
+        print("Dropped '" + n + "' database from MongoDB.")
+        time.sleep(1)
+    if rm == True:
+        if os.path.exists(savesPath + "/" + n):
+            shutil.rmtree(savesPath + "/" + n)
+        if not os.path.exists(savesPath + "/" + n) and eng.DEBUG == 1:
+            print("Removed '" + savesPath + "/" + n + "' from filesystem.")
+            time.sleep(1)
 
 # set a new location
 def setLocation(n):
@@ -34,6 +46,10 @@ def setLocation(n):
     player.update_one( { "SECTION": "location" }, { "$set": { "NAME": n } } )
     locationInfo = rooms.find_one( { "NAME": n } )
     location = locationInfo["NAME"]
+    if eng.DEBUG == 1:
+        print("New location name: " + location)
+        print(locationInfo)
+        time.sleep(1)
 
 # equip a new weapon
 def setWeapon(n):
@@ -42,6 +58,10 @@ def setWeapon(n):
     player.update_one( { "SECTION": "equipped" }, { "$set": { "WEAPON": n } } )
     equippedWeapon = player.find_one( { "SECTION": "equipped" } )["WEAPON"]
     weaponInfo = items.find_one( {"NAME": equippedWeapon } )
+    if eng.DEBUG == 1:
+        print("New weapon name: " + equippedWeapon)
+        print(weaponInfo)
+        time.sleep(1)
 
 # equip a new FX pedal
 def setFX(n):
@@ -50,6 +70,10 @@ def setFX(n):
     player.update_one( { "SECTION": "equipped" }, { "$set": { "FX": n } } )
     addedFX = player.find_one( { "SECTION": "equipped" } )["FX"]
     fxInfo = items.find_one( {"NAME": addedFX } )
+    if eng.DEBUG == 1:
+        print("New FX name: " + addedFX)
+        print(fxInfo)
+        time.sleep(1)
 
 def getInventory():
     global playerInv
@@ -63,14 +87,29 @@ def updateGround(item, action):
     if action == "del":
         # remove item from ground
         rooms.update_one( { "NAME": location }, { "$pull": { "GROUND": item } } )
+        groundList = list(rooms.find_one( { "NAME": location } )["GROUND"])
+        if item in groundList:
+            raise Exception("'" + item + "' still exists on the ground.")
+            return
+        if eng.DEBUG == 1:
+            print("Item to be removed: " + item)
+            print("New GROUND list:")
+            print(groundList)
     elif action == "add":
         # add item to the ground
         groundTemp = list(locationInfo["GROUND"])
         groundTemp.append(item)
         rooms.update_one( { "NAME": location }, { "$set": { "GROUND": groundTemp } } )
+        groundList = list(rooms.find_one( { "NAME": location } )["GROUND"])
+        if item not in groundList:
+            raise Exception("'" + item + "' doesn't exist on the ground.")
+            return
+        if eng.DEBUG == 1:
+            print("Item to be added: " + item)
+            print("New GROUND list:")
+            print(groundList)
     else:
-        print("BUG: Someone forgot to specify an action!")
-        return
+        print("FUNCTION CALL BUG: Someone forgot to specify an action!")
     setLocation(location)
 
 def updateInv(item, action):
@@ -86,19 +125,34 @@ def updateInv(item, action):
             k.remove(item)
         player.update_one( { "SECTION": "inventory" }, { "$set": { "ITEMS": i } } )
         player.update_one( { "SECTION": "inventory" }, { "$set": { "KEY_ITEMS": k } } )
+        getInventory()
+        if item in list(playerInv["ITEMS"]) or item in list(playerInv["KEY_ITEMS"]):
+            raise Exception("'" + item + "' is sill in Ted's inventory.")
+            return
     elif action == "add":
         # add item
         if items.find_one( { "NAME": item } )["TYPE"] == "key":
             k = list(playerInv["KEY_ITEMS"])
             k.append(item)
             player.update_one( { "SECTION": "inventory" }, { "$set": { "KEY_ITEMS": k } } )
+            getInventory()
+            if item not in list(playerInv["KEY_ITEMS"]):
+                raise Exception("'" + item + "' is not in Ted's inventory.")
+                return
         else:
             i = list(playerInv["ITEMS"])
             i.append(item)
             player.update_one( { "SECTION": "inventory" }, { "$set": { "ITEMS": i } } )
+            getInventory()
+            if item not in list(playerInv["ITEMS"]):
+                raise Exception("'" + item + "' is not in Ted's inventory.")
+                return
     else:
-        print("BUG: Someone forgot to specify an action!")
-    getInventory()
+        print("FUNCTION CALL BUG: Someone forgot to specify an action!")
+    if eng.DEBUG == 1:
+        print("New inventory list:")
+        print("Items: " + list(playerInv["ITEMS"]))
+        print("Key Items: " + list(playerInv["KEY_ITEMS"]))
 
 def updateStat(stat, num, action):
     global playerStats
@@ -111,10 +165,12 @@ def updateStat(stat, num, action):
         # decrease the stat
         temp -= int(num)
     else:
-        print("BUG: Someone forgot to specify an action!")
+        print("FUNCTION CALL BUG: Someone forgot to specify an action!")
         return
     player.update_one( { "SECTION": "stats" }, { "$set": { stat : temp } } )
     getStats()
+    if eng.DEBUG == 1:
+        print("New value for " + stat + ": " + playerStats[stat])
 
 # function to set paths and collections
 def define(n):
@@ -145,6 +201,8 @@ def define(n):
     player = db['player']
     # set SLOT_NAME variable
     eng.SLOT_NAME = n
+    if eng.DEBUG == 1:
+        print("Defined all database paths for the game to '" + n + "'")
 
 # save current game state to the database
 def saveGame():
@@ -153,6 +211,8 @@ def saveGame():
         col = db[n]
         # save to a new state file set
         newFile = playerPath + "/" + n + ".json"
+        if eng.DEBUG == 1:
+            print("JSON file path: " + newFile)
         if os.path.exists(newFile):
             os.remove(newFile)
         else:
@@ -167,11 +227,18 @@ def saveGame():
             content = f.read()
             content_new = re.sub('(\{"\$oid": )("[A-Za-z0-9]{24}")(\})', r'\2', content, flags = re.M)
             f.close()
+            if eng.DEBUG == 1:
+                print("Converted BSON to JSON")
         with open(newFile, 'w') as f:
             f.close()
+            if eng.DEBUG == 1:
+                print("Emptied file")
         with open(newFile, 'w') as f:
             f.write(content_new)
             f.close()
+            if eng.DEBUG == 1:
+                print("Saved new JSON to file")
+                time.sleep(1)
 
 # load game state from database
 def loadGame(name):
@@ -192,22 +259,25 @@ def loadGame(name):
         if isinstance(file_data, list):
             col.insert_many(file_data)
             if eng.DEBUG == 1:
-                print("inserted init document to " + n)
+                print("inserted saved document to " + n)
         else:
             col.insert_one(file_data)
             if eng.DEBUG == 1:
-                print("inserted init documents to " + n)
+                print("inserted saved documents to " + n)
     equipment = player.find_one( { "SECTION": "equipped" } )
     setWeapon(equipment["WEAPON"])
     setFX(equipment["FX"])
     locationName = player.find_one( { "SECTION": "location" } )["NAME"]
     setLocation(locationName)
+    if eng.DEBUG == 1:
+        print("Save game loaded: '" + name + "'")
+        time.sleep(1)
 
 # create new database and set initial values
 def newGame(name):
     global equipment
     # remove old save data
-    deleteSave(name)
+    deleteSave(name, True)
     # setup new save data
     define(name)
     # iterate through collections to insert and dump a new save
@@ -229,8 +299,12 @@ def newGame(name):
                 print("inserted init documents to " + n)
         # save to a new state file set
         newFile = playerPath + "/" + n + ".json"
+        if eng.DEBUG == 1:
+            print("JSON file path: " + newFile)
         if os.path.exists(newFile):
             os.remove(newFile)
+            if eng.DEBUG == 1:
+                print("Removed old save game files")
         else:
             pass
         cursor = col.find({})
@@ -243,13 +317,21 @@ def newGame(name):
             content = f.read()
             content_new = re.sub('(\{"\$oid": )("[A-Za-z0-9]{24}")(\})', r'\2', content, flags = re.M)
             f.close()
+            if eng.DEBUG == 1:
+                print("Converted BSON to JSON")
         with open(newFile, 'w') as f:
             f.close()
+            if eng.DEBUG == 1:
+                print("Emptied file")
         with open(newFile, 'w') as f:
             f.write(content_new)
             f.close()
+            if eng.DEBUG == 1:
+                print("Save new JSON to file")
     equipment = player.find_one( { "SECTION": "equipped" } )
     setWeapon(equipment["WEAPON"])
     setFX(equipment["FX"])
     locationName = player.find_one( { "SECTION": "location" } )["NAME"]
     setLocation(locationName)
+    if eng.DEBUG == 1:
+        print("New game created: '" + name + "'")
